move statics into .cc file
Move everything static that’s safe to be included in multiple source
files into a separate .cc file.
