+++ node.gyp
@@ -172,6 +172,7 @@
         'src/node_contextify.cc',
         'src/node_debug_options.cc',
         'src/node_file.cc',
+        'src/node-http2-core.cc',
         'src/node_http2.cc',
         'src/node_http_parser.cc',
         'src/node_javascript.cc',
